requirements.txt
pip==22.0.4
PyGithub==1.55
python-dotenv==0.20.0
requests==2.27.1
